services/api_observations_producer/requirements.txt
kafka-python
pyyaml
httpx
aiolimiter